Second batch of og/newScreen headers
one more batch to go

+++ include/P2JME/Analyzer.h
@@ -9,16 +9,16 @@ namespace P2JME {
 struct Analyzer : public TRenderingProcessorBase {
 	Analyzer(const JMessage::TReference*);
 
-	virtual ~Analyzer() { }                                               // _08 (weak)
-	virtual void do_character(int);                                       // _10
-	virtual void tagColor(const void*, unsigned long);                    // _48 (weak)
-	virtual void tagSize(const void*, unsigned long);                     // _4C (weak)
-	virtual void tagRuby(const void*, unsigned long);                     // _50 (weak)
-	virtual void tagFont(const void*, unsigned long);                     // _54 (weak)
-	virtual void tagImage(unsigned short, const void*, unsigned long);    // _58 (weak)
-	virtual void tagColorEX(unsigned short, const void*, unsigned long);  // _5C (weak)
-	virtual void tagControl(unsigned short, const void*, unsigned long);  // _60 (weak)
-	virtual void tagPosition(unsigned short, const void*, unsigned long); // _64 (weak)
+	virtual ~Analyzer() { }                          // _08 (weak)
+	virtual void do_character(int);                  // _10
+	virtual void tagColor(const void*, u32);         // _48 (weak)
+	virtual void tagSize(const void*, u32);          // _4C (weak)
+	virtual void tagRuby(const void*, u32);          // _50 (weak)
+	virtual void tagFont(const void*, u32);          // _54 (weak)
+	virtual void tagImage(u16, const void*, u32);    // _58 (weak)
+	virtual void tagColorEX(u16, const void*, u32);  // _5C (weak)
+	virtual void tagControl(u16, const void*, u32);  // _60 (weak)
+	virtual void tagPosition(u16, const void*, u32); // _64 (weak)
 
 	void exec(char*);
 
@@ -27,6 +27,9 @@ struct Analyzer : public TRenderingProcessorBase {
 		*out1 = in1;
 		*out2 = in2;
 	}
+
+	// _00     = VTBL
+	// _00-_38 = TRenderingProcessorBase
 };
 
 } // namespace P2JME
